+++ src/PubNub/Models/Consumer/FileSharing/PNPublishFileMessageResult.php
@@ -5,6 +5,8 @@ namespace PubNub\Models\Consumer\FileSharing;
 class PNPublishFileMessageResult
 {
     protected $timestamp;
+    protected ?string $fileId;
+    protected ?string $fileName;
 
     public function __construct($json)
     {
@@ -19,5 +21,27 @@ class PNPublishFileMessageResult
     public function getTimestamp()
     {
         return $this->timestamp;
+    }
+
+    public function getFileId(): ?string
+    {
+        return $this->fileId;
+    }
+
+    public function getFileName(): ?string
+    {
+        return $this->fileName;
+    }
+
+    public function setFileId(string $fileId): self
+    {
+        $this->fileId = $fileId;
+        return $this;
+    }
+
+    public function setFileName(string $fileName): self
+    {
+        $this->fileName = $fileName;
+        return $this;
     }
 }